Re-used VcDetails component in popup preview and normal preview
