+++ src/App.js
@@ -6,7 +6,7 @@ import Register from './pages/Register';
 import Profile from './pages/Profile';
 import VcList from './pages/VcList';
 import FetchVcs from './pages/fetchVcs';
-import VcDetails from './pages/VcDetails';
+import VcDetailsPage from './pages/VcDetails';
 import QrScanner from './pages/QrScanner';
 import Layout from './components/Layout';
 
@@ -43,7 +43,7 @@ const AppRoutes = () => {
         <Route index element={<Navigate to="/vcs" />} />
         <Route path="vcs" element={<VcList />} />
         <Route path="fetch-vcs" element={<FetchVcs />} />
-        <Route path="vcs/:vcId" element={<VcDetails />} />
+        <Route path="vcs/:vcId" element={<VcDetailsPage />} />
         <Route path="profile" element={<Profile />} />
         <Route path="qr-scanner" element={<QrScanner />} />
       </Route>
